Add docstring for 'cache-validity'
We remain hacky.
To my mind, this is something where you explore the types of info
you need and where you need them and once you got a decent idea of
that, you can sit down and  sketch out efficient methods of getting
and maybe holding on to the info.